% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Better_vlnplot.R
\name{Better_vlnplot}
\alias{Better_vlnplot}
\title{Generate violin plots from a Seurat object and vector of genes.}
\usage{
Better_vlnplot(
  seurat,
  clus_ident = "seurat_clusters",
  condition_ident = vector(),
  features,
  assay = "RNA",
  idents_to_plot = vector(),
  ncol = 1
)
}
\arguments{
\item{seurat}{seurat object}

\item{clus_ident}{Identity for clusters. 'seurat_clusters' by default.}

\item{condition_ident}{name of metadata column to split plots by}

\item{features}{vector of gene/feature names to plot}

\item{assay}{assay to plot from. default = 'RNA'}

\item{idents_to_plot}{vector of a subset of specific idents within clus_ident to plot}

\item{ncol}{number of columns if plotting multiple genes}
}
\description{
Generate violin plots from a Seurat object and vector of genes.
}
