% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetterVlnPlot.R
\name{BetterVlnPlot}
\alias{BetterVlnPlot}
\title{Generate violin plots from a Seurat object and vector of genes.}
\usage{
BetterVlnPlot(
  seurat,
  clus_ident = "seurat_clusters",
  condition_ident = vector(),
  features,
  assay = "RNA",
  idents_to_plot = vector(),
  ncol = 1,
  dot_size = 1,
  y_axis_title = "log normalized counts"
)
}
\arguments{
\item{seurat}{seurat object}

\item{clus_ident}{Identity for clusters. 'seurat_clusters' by default.}

\item{condition_ident}{name of metadata column to split plots by}

\item{features}{vector of gene/feature names to plot}

\item{assay}{assay to plot from. default = 'RNA'}

\item{idents_to_plot}{vector of a subset of specific idents within clus_ident to plot}

\item{ncol}{number of columns if plotting multiple genes}

\item{dot_size}{size of dots. Set to NA to remove dots}

\item{y_axis_title}{y-axis label. Default is 'log normalized counts'}
}
\description{
Generate violin plots from a Seurat object and vector of genes.
}
