% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportSeuratMeta.R
\name{ExportSeuratMeta}
\alias{ExportSeuratMeta}
\title{Export meta data files for RNA velocity and CellRank analysis}
\usage{
ExportSeuratMeta(seurat, clus_ident, sample_ident, dir)
}
\arguments{
\item{seurat}{Path to Seurat .rds file}

\item{clus_ident}{Identity for clusters. Normally 'seurat_clusters' but can be any identity. I generally use 'annotations' to store cluster annotations}

\item{sample_ident}{Identity for samples. Genereally 'orig.ident'}

\item{dir}{Directory to output .csv files}
}
\value{
.csv files with cell by embeddings and clusters. .csv file with color palette for plotting in scVelo, CellRank, etc.
}
\description{
Export meta data files for RNA velocity and CellRank analysis
}
