% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindClusterSweep.R
\name{FindClusterSweep}
\alias{FindClusterSweep}
\title{Run common single-cell RNA-seq clustering algorithms as implemented in Seurat across a range of resolution values and compute common clustering metrics. This function assumes a KNN graph already exists in the specified assay. Run seurat's FindNeighbors before this function.}
\usage{
FindClusterSweep(
  seurat,
  assay = "RNA",
  resolutions = c(0.2, 0.4, 0.6, 0.8, 1, 1.2, 1.4, 1.6),
  algorithm = 1,
  conda,
  pca_dim = c(1:30),
  reduction = "pca",
  plot_reduction = "umap",
  file_name = "FindClusterSweep_plots"
)
}
\arguments{
\item{seurat}{A seurat object}

\item{assay}{seurat assay e.g. 'RNA'}

\item{resolutions}{Vector of clustering resolutions}

\item{algorithm}{Seurat FindClusters algorithm parameter. From Seurat: 'Algorithm for modularity optimization 1 = original Louvain algorithm; 2 = Louvain algorithm with multilevel refinement; 3 = SLM algorithm; 4 = Leiden algorithm. Leiden requires the leidenalg python.'}

\item{conda}{If applicable, path to conda environment. Required for Leiden algorithm = 4}

\item{pca_dim}{vector of dimensions to use for computing silheoutte scores. default = 0,30}

\item{reduction}{reduction to use for computing silhouette scores. default = 'pca'}

\item{plot_reduction}{reduction to plot silhouette scores. default = 'umap'}

\item{file_name}{plot .pdf file name. default = 'FindClusterSweep_plots'}
}
\value{
A seurat object with clustering. .pdf document with a series of clustering related plots
}
\description{
Run common single-cell RNA-seq clustering algorithms as implemented in Seurat across a range of resolution values and compute common clustering metrics. This function assumes a KNN graph already exists in the specified assay. Run seurat's FindNeighbors before this function.
}
