% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_assay_data_compat}
\alias{get_assay_data_compat}
\title{Compatibility wrapper for Seurat assay data access}
\usage{
get_assay_data_compat(seurat, assay, slot)
}
\description{
Compatibility wrapper for Seurat assay data access
}
\keyword{internal}
