% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_pheatmap}
\alias{safe_pheatmap}
\title{Plot heatmaps safely when input values are constant}
\usage{
safe_pheatmap(mat, ...)
}
\description{
Plot heatmaps safely when input values are constant
}
\keyword{internal}
