% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedVlnPlot.R
\name{StackedVlnPlot}
\alias{StackedVlnPlot}
\title{Generate a stacked violin plot from a Seurat object and list of genes.}
\usage{
StackedVlnPlot(
  seurat,
  clus_ident = "seurat_clusters",
  features,
  assay = "RNA",
  slot = "data"
)
}
\arguments{
\item{seurat}{A Seurat object}

\item{clus_ident}{Identity for clusters. 'seurat_clusters' by default.}

\item{features}{vector of gene names to plot}

\item{assay}{assay to plot from. default = 'RNA'}

\item{slot}{vslot to plot from. default = 'data'}
}
\description{
Generate a stacked violin plot from a Seurat object and list of genes.
}
