% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateAnno.R
\name{UpdateAnno}
\alias{UpdateAnno}
\title{Update cell annotations.}
\usage{
UpdateAnno(seurat, old_anno_ident, new_anno_ident, convert_file)
}
\arguments{
\item{seurat}{Seurat object}

\item{old_anno_ident}{Identity of annotations to be changed}

\item{new_anno_ident}{Identity for new annotations}

\item{convert_file}{absolute path to .csv file containing 2 columns. 1st column should have the header identical to \code{old_anno_ident} and contain each annotation to update as a row. The 2nd column should have a header identical to \code{new_anno_ident} and contain the updated annotations matched to \code{old_anno_ident}}
}
\value{
Seurat object with updated cell annotations specified in \code{convert_file}
}
\description{
Update cell annotations.
}
