% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindMarkers.R
\name{FindMarkers}
\alias{FindMarkers}
\title{Differential expression testing between 2 specified groups of cells via pseudobulking and DESeq2}
\usage{
FindMarkers(
  seurat,
  clus_ident,
  group_1,
  group_2,
  sample_ident,
  expfilt_counts = 1,
  expfilt_freq = 0.5,
  out_dir = "FindMarkers",
  alpha = 0.1,
  assay = "RNA"
)
}
\arguments{
\item{seurat}{A Seurat object}

\item{clus_ident}{Identity for clusters. Normally 'seurat_clusters' but can be any identity}

\item{group_1}{Identity of cells within \code{clus_ident} to compare}

\item{group_2}{Identity of cells within \code{clus_ident} to compare}

\item{sample_ident}{Sample identities. Identity class that indicates how to partition samples}

\item{expfilt_counts}{genes with less than \code{expfilt_counts} in \code{expfilt_freq * sample number} will be removed from DESeq2 model. 1 by default.}

\item{expfilt_freq}{genes that have greater than \code{expfilt_counts} in greater than \code{expfilt_freq} fraction of cells will be kept for the DESeq2 model. 0.5 by default}

\item{out_dir}{Name of output directory}

\item{alpha}{FDR adjusted p-value threshold for significance in plotting. 0.1 by default.}

\item{assay}{Which assay to use. RNA by default. I added this parameter to enable use of ADT data when desired.}
}
\value{
.csv files with marker genes per \code{clus_ident}. .pdf files with diagnostic plots
}
\description{
Differential expression testing between 2 specified groups of cells via pseudobulking and DESeq2
}
