% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindMarkersBulk.R
\name{FindMarkersBulk}
\alias{FindMarkersBulk}
\title{Find all markers via pseudobulking and DESeq2}
\usage{
FindMarkersBulk(
  seurat,
  clus_ident,
  sample_ident,
  expfilt_counts = 1,
  expfilt_freq = 0.5,
  n_top_genes = 50,
  pct.in = 25,
  out_dir = "FindMarkersBulk_outs",
  alpha = 0.1
)
}
\arguments{
\item{seurat}{A Seurat object}

\item{clus_ident}{Identity for clusters. Normally 'seurat_clusters' but can be any identity}

\item{sample_ident}{Sample identities. Identity class that indicates how to partition samples}

\item{expfilt_counts}{genes with less than \code{expfilt_counts} in \code{expfilt_freq * sample number} will be removed from DESeq2 mode. 1 by default.}

\item{expfilt_freq}{genes that have greater than \code{expfilt_counts} in greater than \code{expfilt_freq} fraction of cells will be kept for the DESeq2 model. 0.5 by default}

\item{n_top_genes}{number of top genes per cluster to save and make a heatmap with}

\item{pct.in}{Filter threshold for top marker genes per cluster. For a given gene and cluster, if the fraction of cells with counts is less than pct.in it is removed from top_markers.}

\item{out_dir}{Name of output directory}

\item{alpha}{FDR adjusted p-value threshold for significance in plotting. 0.1 by default.}
}
\value{
.csv files with marker genes per \code{clus_ident}. .pdf files with plots
}
\description{
Find all markers via pseudobulking and DESeq2
}
