% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindMarkersCondition.R
\name{FindMarkersCondition}
\alias{FindMarkersCondition}
\title{Differential expression testing between conditions via pseudobulking and DESeq2}
\usage{
FindMarkersCondition(
  seurat,
  clus_ident,
  sample_ident,
  condition_ident,
  conditions,
  expfilt_counts = 1,
  expfilt_freq = 0.5,
  out_dir = "FindMarkersCondition_outs",
  alpha = 0.1,
  assay = "RNA"
)
}
\arguments{
\item{seurat}{A Seurat object}

\item{clus_ident}{Identity for clusters. Normally 'seurat_clusters' but can be any identity}

\item{sample_ident}{Sample identities. Identity class that indicates how to partition samples}

\item{condition_ident}{Identity class for conditions to be tested}

\item{conditions}{A vector of the 2 conditions within \code{condition_ident} to be included in the DESeq2 model.}

\item{expfilt_counts}{genes with less than \code{expfilt_counts} in \code{expfilt_freq * sample number} will be removed from DESeq2 model. 1 by default.}

\item{expfilt_freq}{genes that have greater than \code{expfilt_counts} in greater than \code{expfilt_freq} fraction of cells will be kept for the DESeq2 model. 0.5 by default}

\item{out_dir}{Name of output directory}

\item{alpha}{FDR adjusted p-value threshold for significance in plotting. 0.1 by default.}

\item{assay}{Which assay to use. RNA by default. I added this parameter to enable use of ADT data when desired.}
}
\value{
.csv files with marker genes per \code{clus_ident}. .pdf files with diagnostic plots
}
\description{
Differential expression testing between conditions via pseudobulking and DESeq2
}
