% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindMarkersCondition.R
\name{FindMarkersCondition}
\alias{FindMarkersCondition}
\title{Differential expression testing between conditions via pseudobulking and DESeq2}
\usage{
FindMarkersCondition(
  seurat,
  clus_ident,
  sample_ident,
  condition_ident,
  conditions,
  expfilt
)
}
\arguments{
\item{seurat}{A Seurat object}

\item{clus_ident}{Identity for clusters. Normally 'seurat_clusters' but can be any identity}

\item{sample_ident}{Sample identities. Identity class that indicates how to partition samples}

\item{condition_ident}{Identity class for conditions to be tested}

\item{conditions}{A vector of the 2 conditions within condition_ident to be included in the DESeq2 model.}

\item{expfilt}{genes that have greater than 0 counts in greater than expfilt fraction of cells will be kept for the DESeq2 model}
}
\value{
.csv files with marker genes per clus_ident. .pdf files with plots
}
\description{
Differential expression testing between conditions via pseudobulking and DESeq2
}
